Cleaned up & simplified code a ton.
