+++ DriveBackup/src/main/java/ratismal/drivebackup/DriveBackupApi.java
@@ -1,8 +1,5 @@
 package ratismal.drivebackup;
 
-import org.bukkit.Bukkit;
-import org.bukkit.scheduler.BukkitScheduler;
-
 import java.util.ArrayList;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
@@ -10,6 +7,7 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
 
+import ratismal.drivebackup.plugin.DriveBackup;
 import ratismal.drivebackup.util.MessageUtil;
 
 public class DriveBackupApi {
@@ -113,9 +111,6 @@ public class DriveBackupApi {
      */
     public static void reloadConfig() {
         DriveBackup.reloadLocalConfig();
-        BukkitScheduler scheduler = Bukkit.getServer().getScheduler();
-        scheduler.cancelTasks(DriveBackup.getInstance());
-        DriveBackup.startThread();
     }
 
     /**
