+++ DriveBackup/src/main/java/ratismal/drivebackup/handler/commandHandler/CommandInfo.java
@@ -1,0 +1,5 @@
+package ratismal.drivebackup.handler.commandHandler;
+
+class CommandInfo {
+    
+}
